GL 3.2 - no crashes but no rendering either

+++ java_src/HeightField.java
@@ -10,17 +10,17 @@ import java.lang.Math;
 import javax.media.jai.InterpolationBilinear;
 
 class HeightField {
-  // The height data is stored in a 2d array of doubles.
+  // The height data is stored in a 2d array of floats.
   //
   // To map from the array to world coordinates, the proper X, Y, and Z scale factors must be known. 
   // Scale is defined as world units per height field "pixel."
   // squareSize is the size in world coords of the mesh squares.
   int cols, rows;
-  public double xScale, yScale, zScale; // cols => x, rows => y, height => z
-  double heights[][];
+  public float xScale, yScale, zScale; // cols => x, rows => y, height => z
+  float heights[][];
   
   // Can gracefully handle out-of-range coordinates
-  public double atXY(double worldX, double worldY) {
+  public float atXY(float worldX, float worldY) {
     // scale * height array offset = world coord
     // height array offset = world coord / scale
     int left   = Math.max((int)Math.floor(worldX / xScale), 0   );
@@ -37,7 +37,7 @@ class HeightField {
     ) * zScale;
   }
   
-  public HeightField(double xScale, double yScale, double zScale) {
+  public HeightField(float xScale, float yScale, float zScale) {
     this.xScale = xScale;
     this.yScale = yScale;
     this.zScale = zScale;
@@ -49,7 +49,7 @@ class HeightField {
     BufferedImage img = javax.imageio.ImageIO.read(new java.io.File(imagePath));
     cols = img.getWidth();
     rows = img.getHeight();
-    heights = new double[cols][rows];
+    heights = new float[cols][rows];
     for (int y = 0; y < img.getHeight(); y++) {
       for (int x = 0; x < img.getWidth(); x++) {
         heights[x][y] = (new Color(img.getRGB(x, y))).getRed();
@@ -57,11 +57,11 @@ class HeightField {
     }
   }
   
-  public double worldLength() {
+  public float worldLength() {
     return cols * yScale;
   }
   
-  public double worldWidth() {
+  public float worldWidth() {
     return rows * xScale;
   }
 }
